// src/app/models/dates.interface.ts
export interface Dates {
  id?: string;
  nombre: string;
  apellidos: string;
  correo: string;
  telefono: string;
  fecha: string;
  personas: string;
  tipoEvento: string;
  alberca: string;
  horaInicio: number;
  horasExtras: number;
  horaFin: number;
  colorMantel: string;
  extras: string[];
}